extract gesture, zoom, and pinch logic from GalaxyMapRender into hooks
GalaxyMapRender was ~500 lines mixing gesture prevention, zoom/pan, touch
pinch, and rendering. Extracts three focused hooks:

- usePreventBrowserZoom: blocks native browser zoom gestures, tracks resize
- useZoomPan: wheel zoom and drag-to-pan with throttling
- usePinchZoom: touch pinch-to-zoom with jitter prevention

Shared helpers (getDistance, requestBatchDraw) moved to canvasUtils.ts.
GalaxyMapRender is now ~267 lines focused on rendering and state wiring.

Adds tests/galaxymap-hooks.test.ts (20 tests) verifying hook structure,
exports, shared util usage, and GalaxyMap integration.
