+++ src/components/hooks/canvasUtils.ts
@@ -1,0 +1,18 @@
+import Konva from 'konva';
+
+export function getDistance(touch1: Touch, touch2: Touch): number {
+  const dx = touch1.clientX - touch2.clientX;
+  const dy = touch1.clientY - touch2.clientY;
+  return Math.sqrt(dx * dx + dy * dy);
+}
+
+let frameRequested = false;
+export function requestBatchDraw(stage: Konva.Stage): void {
+  if (!frameRequested) {
+    frameRequested = true;
+    requestAnimationFrame(() => {
+      stage.batchDraw();
+      frameRequested = false;
+    });
+  }
+}
